fix formatting issues + move stuff around
